feat: Добавил поле useful в таблицу film_review в БД. Изменил структу в стиле исходного проекта. Сменил Long на int у FilmReview. Добавил валидацию полей. Покрыл все интеграционными тестами.

+++ src/main/java/ru/yandex/practicum/filmorate/model/ReviewLike.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ReviewLike {
-    private Long reviewId;
+    private int reviewId;
     private int userId;
     private boolean positive;
 }
